Wide metal spacing klayout: zero space is fine
In other words, the metals are shrunk then put back at the same size (to
eliminate features that are not wide).
Then the separation of wide_metal to the metal is made.

There can however be an issue when edges of wide metal and regular metal ends
up being aligned, as this will be detected as an error when there is actually
none, as the metals are touching.

Fixes #18.